All functions should pass all the tests and norminette

+++ ft_memchr.c
@@ -6,22 +6,22 @@
 /*   By: mtravez <mtravez@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 15:10:48 by mtravez           #+#    #+#             */
-/*   Updated: 2022/10/15 18:44:14 by mtravez          ###   ########.fr       */
+/*   Updated: 2022/10/18 16:31:06 by mtravez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 //This function searches within the first n bytes of s
-//for the first occurrance of c
-void	*ft_memchr(const void *s, int c, unsigned int n)
+//for the first occurrance of c and returns the pointer to it
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	int					i;
 	const unsigned char	*temp;
 
 	i = 0;
 	temp = (const unsigned char *) s;
-	while ((unsigned int)i < n)
+	while ((size_t)i < n)
 	{
 		if (temp[i] == (unsigned char)c)
 			return ((void *)&temp[i]);
